refactor(components): [carousel] refactor
