+++ packages/components/carousel/src/carousel-item.ts
@@ -1,0 +1,15 @@
+import { buildProps } from '@element-plus/utils'
+import type { ExtractPropTypes } from 'vue'
+import type CarouselItem from './carousel-item.vue'
+
+export const carouselItemProps = buildProps({
+  name: { type: String, default: '' },
+  label: {
+    type: [String, Number],
+    default: '',
+  },
+} as const)
+
+export type CarouselItemProps = ExtractPropTypes<typeof carouselItemProps>
+
+export type CarouselItemInstance = InstanceType<typeof CarouselItem>
