+++ packages/components/carousel/src/carousel.ts
@@ -1,51 +1,57 @@
-import type { Ref, ToRefs, UnwrapRef } from 'vue'
+import { buildProps, isNumber } from '@element-plus/utils'
+import type { ExtractPropTypes } from 'vue'
+import type Carousel from './carousel.vue'
 
-export interface ICarouselItemProps {
-  name: string
-  label: string | number
-  key: string
+export const carouselProps = buildProps({
+  initialIndex: {
+    type: Number,
+    default: 0,
+  },
+  height: { type: String, default: '' },
+  trigger: {
+    type: String,
+    default: 'hover',
+  },
+  autoplay: {
+    type: Boolean,
+    default: true,
+  },
+  interval: {
+    type: Number,
+    default: 3000,
+  },
+  indicatorPosition: { type: String, default: '' },
+  indicator: {
+    type: Boolean,
+    default: true,
+  },
+  arrow: {
+    type: String,
+    default: 'hover',
+  },
+  type: { type: String, default: '' },
+  loop: {
+    type: Boolean,
+    default: true,
+  },
+  direction: {
+    type: String,
+    default: 'horizontal',
+    validator(val: string) {
+      return ['horizontal', 'vertical'].includes(val)
+    },
+  },
+  pauseOnHover: {
+    type: Boolean,
+    default: true,
+  },
+} as const)
+
+export const carouselEmits = {
+  change: (current: number, prev: number) => [current, prev].every(isNumber),
 }
 
-export interface ICarouselItemData {
-  hover: boolean
-  translate: number
-  scale: number
-  active: boolean
-  ready: boolean
-  inStage: boolean
-  animating: boolean
-}
+export type CarouselProps = ExtractPropTypes<typeof carouselProps>
+export type CarouselEmits = typeof carouselEmits
 
-export interface ICarouselProps {
-  initialIndex: number
-  height: string
-  trigger: string
-  autoplay: boolean
-  interval: number
-  indicatorPosition: string
-  indicator: boolean
-  arrow: string
-  type: string
-  loop: boolean
-  direction: string
-  pauseOnHover: boolean
-}
-
-export type UnionCarouselItemData = ICarouselItemProps &
-  ToRefs<ICarouselItemData>
-
-export interface CarouselItem extends UnionCarouselItemData {
-  uid: number
-  translateItem: (index: number, activeIndex: number, oldIndex: number) => void
-}
-
-export interface InjectCarouselScope {
-  root: Ref<HTMLElement>
-  direction: string
-  type: string
-  items: Ref<UnwrapRef<CarouselItem[]>>
-  loop: boolean
-  addItem: (item: CarouselItem) => void
-  removeItem: (uid: number) => void
-  setActiveItem: (index: number) => void
-}
+export type CarouselInstance = InstanceType<typeof Carousel>
